+++ aitune/torch/backend/__init__.py
@@ -3,6 +3,7 @@
 """Torch backend module."""
 
 from aitune.torch.backend.backend import Backend, BuildMode, ExecutionMode
+from aitune.torch.backend.kernel_optimizer_backend import KernelOptimizerBackend, KernelOptimizerBackendConfig
 from aitune.torch.backend.onnx_runtime_backend import (
     ONNXExecutionProvider,
     ONNXRuntimeBackend,
@@ -36,6 +37,8 @@ __all__ = [
     "Backend",
     "BuildMode",
     "ExecutionMode",
+    "KernelOptimizerBackend",
+    "KernelOptimizerBackendConfig",
     "ONNXAutoCastConfig",
     "ONNXExecutionProvider",
     "ONNXQuantizationConfig",